pyproject.toml
[project]
name = "agent-python"
version = "0.1.0"
authors = [
    { name = "Mohid Yousaf", email = "mohid.yousaf@whizzbridge.com" }
]
requires-python = ">=3.12,<3.13"
dependencies = [
    "cartesia>=2.0.15",
    "elevenlabs>=2.22.0",
    "email-validator>=2.0.0",
    "livekit-agents[assemblyai,deepgram,openai]~=1.2.14",
    "livekit-plugins-cartesia~=1.2.0",
    "livekit-plugins-elevenlabs~=1.2.17",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-turn-detector~=1.2.11",
    "logger>=1.4",
    "openai>=1.0.0",
    "pyaudio>=0.2.14",
    "pydantic[email]>=2.0.0",
    "pyjwt>=2.0.0",
    "python-dotenv>=0.19.0",
    "sqlalchemy[asyncio]>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
]

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"
